% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ogre_sampleInfo.R
\name{checkLimsNumbers}
\alias{checkLimsNumbers}
\title{checkLimsNumbers}
\usage{
checkLimsNumbers(x)
}
\arguments{
\item{x}{data frame as returned by \code{\link{getInfoOnAnalysedSamplesFromExcel}}
or \code{\link{getInfoOnAnalysedSamplesForStation}}}
}
\value{
duplicated entries (if any)
}
\description{
check for duplicate LIMS numbers in sample information
}
