% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ogre_UserDefinedFunctions.R
\name{downloadLatestPankeDataFromQuantum}
\alias{downloadLatestPankeDataFromQuantum}
\title{Download Panke Data from Quantum}
\usage{
downloadLatestPankeDataFromQuantum(target.dir)
}
\arguments{
\item{target.dir}{full path to target directory}
}
\value{
full path(s) to downloaded csv file(s)
}
\description{
Download latest Panke data from Quantum web portal
}
