% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ogre_bottleSelection.R
\name{ganttPlotSamples_v1}
\alias{ganttPlotSamples_v1}
\title{Gantt Plot Samples (version 1)}
\usage{
ganttPlotSamples_v1(bottleEvents, endTimeOffsets)
}
\arguments{
\item{bottleEvents}{data frame as returned by \code{\link{samplingEventsToBottleEvents}}, 
with begin and end of time intervals represented by an auto sampler's bottle.}

\item{endTimeOffsets}{vector of integer offsets determining new end times for 
the bottle events by: bottleEvents$tEnd <- bottleEvents$tBeg + 60*endTimeOffsets - 1}
}
\value{
ganttPlotSamples version 1
}
\description{
Gantt Plot Samples (version 1)
}
