% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ogre_bottleSelection.R
\name{ganttPlotSamples_v2}
\alias{ganttPlotSamples_v2}
\title{Gantt Plot Samples (version 2)}
\usage{
ganttPlotSamples_v2(bottleEvents, bottle, duration)
}
\arguments{
\item{bottleEvents}{data frame as returned by \code{\link{samplingEventsToBottleEvents}}, 
with begin and end of time intervals represented by an auto sampler's bottle.}

\item{bottle}{bottle number}

\item{duration}{"duration" of "bottle event"}
}
\value{
ganttPlotSamples version 2
}
\description{
Gantt Plot Samples (version 2)
}
