% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ogre_UserDefinedFunctions.R
\name{getCurrentFlowSubDirectory}
\alias{getCurrentFlowSubDirectory}
\title{getCurrentFlowSubDirectory}
\usage{
getCurrentFlowSubDirectory(flowDirectory, do.stop = TRUE)
}
\arguments{
\item{flowDirectory}{full path to the directory}

\item{do.stop}{if \code{TRUE} (default) the program will stop if there are 
unexpected files or folders, otherwise a warning is given}
}
\value{
path to most recent flow sub-directory
}
\description{
getCurrentFlowSubDirectory
}
