% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ogre_sampleInfo.R
\name{getInfoOnAnalysedSamplesForStation}
\alias{getInfoOnAnalysedSamplesForStation}
\title{getInfoOnAnalysedSamplesForStation}
\usage{
getInfoOnAnalysedSamplesForStation(
  xls,
  station,
  dbg = FALSE,
  sheetPrefix = "data_"
)
}
\arguments{
\item{xls}{full path to Excel file}

\item{station}{three letter code of monitoring station for which data are to be read}

\item{dbg}{if TRUE, debug messages are shown, else not}

\item{sheetPrefix}{prefix of sheet names of sheets to be read
(default: "data_")}
}
\value{
data frame with columns \emph{station}, \emph{LIMS_Nr}, 
\emph{BAK_LIMS_Nr}, \emph{firstSampling}, \emph{lastSampling} and 
\emph{Art_der_Probe}
}
\description{
Read information on analysed samples for one monitoring 
station from Excel file maintained by OGRE team
}
\seealso{
\code{\link{getInfoOnAnalysedSamplesFromExcel}}
}
