% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ogre_sampleInfo.R
\name{getInfoOnAnalysedSamplesFromExcel}
\alias{getInfoOnAnalysedSamplesFromExcel}
\title{getInfoOnAnalysedSamplesFromExcel}
\usage{
getInfoOnAnalysedSamplesFromExcel(xls, dbg = FALSE, sheetPrefix = "data_")
}
\arguments{
\item{xls}{full path to Excel file}

\item{dbg}{if TRUE, debug messages are shown, else not}

\item{sheetPrefix}{prefix of sheet names of sheets to be read
(default: "data_")}
}
\value{
data frame with ...
}
\description{
Read information on analysed samples from Excel file 
maintained by OGRE team
}
\seealso{
\code{\link{getInfoOnAnalysedSamplesForStation}}
}
