% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogreConstants.R
\name{OGRE_DICTIONARY_FILE}
\alias{OGRE_DICTIONARY_FILE}
\title{Default folder "dictionary" file}
\usage{
OGRE_DICTIONARY_FILE()
}
\value{
path to OGRE path dictionary file
}
\description{
default "dictionary" file describing the folder structure to be 
used in OGRE
}
\examples{
OGRE_DICTIONARY_FILE()

}
