% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogreConstants.R
\name{OGRE_ODM_LABMETHODS}
\alias{OGRE_ODM_LABMETHODS}
\title{OGRE_ODM_LABMETHODS}
\usage{
OGRE_ODM_LABMETHODS(stringsAsFactors = default.stringsAsFactors())
}
\arguments{
\item{stringsAsFactors}{TRUE or FALSE (default: \code{\link{default.stringsAsFactors}})
passed on to data.frame()}
}
\value{
data frame with labmethods
}
\description{
OGRE_ODM_LABMETHODS
}
\examples{
str(OGRE_ODM_LABMETHODS())
}
