% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogreConstants.R
\name{OGRE_ODM_METHODS}
\alias{OGRE_ODM_METHODS}
\title{OGRE_ODM_METHODS}
\usage{
OGRE_ODM_METHODS(stringsAsFactors = default.stringsAsFactors())
}
\arguments{
\item{stringsAsFactors}{TRUE or FALSE (default: \code{\link{default.stringsAsFactors}})
passed on to data.frame()}
}
\value{
data frame with methods
}
\description{
OGRE_ODM_METHODS
}
\examples{
str(OGRE_ODM_METHODS())
}
