% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogreConstants.R
\name{OGRE_ODM_SAMPLE_TYPES}
\alias{OGRE_ODM_SAMPLE_TYPES}
\title{OGRE_ODM_SAMPLE_TYPES}
\usage{
OGRE_ODM_SAMPLE_TYPES(stringsAsFactors = default.stringsAsFactors())
}
\arguments{
\item{stringsAsFactors}{TRUE or FALSE (default: \code{\link{default.stringsAsFactors}})
passed on to data.frame()}
}
\value{
data frame with sample types
}
\description{
OGRE_ODM_SAMPLE_TYPES
}
\examples{
str(OGRE_ODM_SAMPLE_TYPES())
}
