% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogreConstants.R
\name{OGRE_ODM_SITES}
\alias{OGRE_ODM_SITES}
\title{OGRE_ODM_SITES}
\usage{
OGRE_ODM_SITES(stringsAsFactors = default.stringsAsFactors())
}
\arguments{
\item{stringsAsFactors}{TRUE or FALSE (default: \code{\link{default.stringsAsFactors}})
passed on to data.frame()}
}
\value{
data frame with sources
}
\description{
TODO: further information on the sites (e.g. addresses need to be
added manually within the database...)
}
\examples{
str(OGRE_ODM_SITES())
}
