% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogreConstants.R
\name{OGRE_ODM_SOURCES}
\alias{OGRE_ODM_SOURCES}
\title{OGRE_ODM_SOURCES}
\usage{
OGRE_ODM_SOURCES(stringsAsFactors = default.stringsAsFactors())
}
\arguments{
\item{stringsAsFactors}{TRUE or FALSE (default: \code{\link{default.stringsAsFactors}})
passed on to data.frame()}
}
\value{
data frame with sources
}
\description{
OGRE_ODM_SOURCES
}
\examples{
str(OGRE_ODM_SOURCES())
}
