% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogreConstants.R
\name{OGRE_ODM_UNITS}
\alias{OGRE_ODM_UNITS}
\title{OGRE_ODM_UNITS}
\usage{
OGRE_ODM_UNITS(stringsAsFactors = default.stringsAsFactors())
}
\arguments{
\item{stringsAsFactors}{TRUE or FALSE (default: \code{\link{default.stringsAsFactors}})
passed on to data.frame()}
}
\value{
data frame with units
}
\description{
OGRE_ODM_UNITS
}
\examples{
str(OGRE_ODM_UNITS())
}
