% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogreConstants.R
\name{OGRE_SITES}
\alias{OGRE_SITES}
\title{OGRE_SITES}
\usage{
OGRE_SITES()
}
\value{
list of lists. One list element per Site, each of which is a list
with exactly one list element: SiteID, holding the SiteID of the
corresponding site
}
\description{
OGRE_SITES
}
