% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogreConstants.R
\name{OGRE_VARIABLES}
\alias{OGRE_VARIABLES}
\title{OGRE_VARIABLES}
\usage{
OGRE_VARIABLES()
}
\value{
data frame with laboratory variables
}
\description{
OGRE_VARIABLES
}
\examples{
str(OGRE_VARIABLES())

}
