% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ogre_bottleSelection.R
\name{openBottleSelector}
\alias{openBottleSelector}
\title{Open Bottle Selecter}
\usage{
openBottleSelector(bottleEvents, version = 2)
}
\arguments{
\item{bottleEvents}{data frame as returned by \code{\link{samplingEventsToBottleEvents}}, 
with begin and end of time intervals represented by an auto sampler's bottle.}

\item{version}{1 or 2 (default: 2), used to switch between different plotting 
options}
}
\value{
opens bottle selector
}
\description{
Open Bottle Selecter
}
