% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogreRead.R
\name{read_BWB_LaboratoryReportFromXls}
\alias{read_BWB_LaboratoryReportFromXls}
\title{read_BWB_LaboratoryReportFromXls}
\usage{
read_BWB_LaboratoryReportFromXls(
  labResult.xls,
  sheetName = "Tabelle1",
  date.format = kwb.utils::underscoreToPercent("_d._m._Y"),
  methodRequired = FALSE,
  dbg = FALSE,
  open.on.error = TRUE
)
}
\arguments{
\item{labResult.xls}{full path to Excel file}

\item{sheetName}{name of sheet in Excel file. Default: "Tabelle1"}

\item{date.format}{date format used in Excel file. Default: "\%d.\%m.\%Y"}

\item{methodRequired}{if TRUE, only those rows of the Excel file are 
considered in which a method is given in column "Methode". Default: FALSE}

\item{dbg}{print debug messages (default: FALSE)}

\item{open.on.error}{open on error (default: TRUE)}
}
\value{
data frame with columns \emph{VariableCode}, \emph{LabSampleCode}, 
\emph{SamplingDate}, \emph{SiteCode}, \emph{DataValueText}, \emph{LabMethodName}, 
\emph{LabUnits}
}
\description{
NOTE:  only rows are used in which parameter, LabMethodName and 
LabUnits are given
}
