% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ogre_UserDefinedFunctions.R
\name{read_hydraulics}
\alias{read_hydraulics}
\title{read_hydraulics}
\usage{
read_hydraulics(
  settings,
  dictionary = kwb.utils::selectElements(settings, "dictionary"),
  do.stop = TRUE
)
}
\arguments{
\item{settings}{list of settings as returned by \code{\link{configure}}}

\item{dictionary}{list of path definitions}

\item{do.stop}{if \code{TRUE} (default) the program will stop if there are 
unexpected files or folders, otherwise a warning is given}
}
\value{
data frame with columns \emph{DateTime} (POSIXct, UTC), \emph{H},
\emph{v}, \emph{Q}, \emph{T}
}
\description{
read_hydraulics
}
