% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ogre_UserDefinedFunctions.R
\name{readOgreSamplerFileByName}
\alias{readOgreSamplerFileByName}
\title{readOgreSamplerFileByName}
\usage{
readOgreSamplerFileByName(samplerFile, bottlesToConsider, siteCode = NA)
}
\arguments{
\item{samplerFile}{full path to sampler file}

\item{bottlesToConsider}{bottlesToConsider}

\item{siteCode}{siteCode (default: NA)}
}
\value{
data frame with ???
}
\description{
readOgreSamplerFileByName
}
