% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ogre_UserDefinedFunctions.R
\name{stopOnWrongSampleSite}
\alias{stopOnWrongSampleSite}
\title{stopOnWrongSampleSite}
\usage{
stopOnWrongSampleSite(sampleData, siteCode, samplerFile = "<unknown>")
}
\arguments{
\item{sampleData}{sampleData}

\item{siteCode}{siteCode}

\item{samplerFile}{samplerFile (default: "<unknown>")}
}
\value{
error if site_id not found
}
\description{
stopOnWrongSampleSite
}
