% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ogre_UserDefinedFunctions.R
\name{usePredictedFlowInTimeInterval}
\alias{usePredictedFlowInTimeInterval}
\title{usePredictedFlowInTimeInterval}
\usage{
usePredictedFlowInTimeInterval(hydraulicData, firstTimestamp, lastTimestamp)
}
\arguments{
\item{hydraulicData}{hydraulicData}

\item{firstTimestamp}{firstTimestamp}

\item{lastTimestamp}{lastTimestamp}
}
\value{
data frame with filtered hydraulic data (first - last timestamp)
}
\description{
usePredictedFlowInTimeInterval
}
